test(e2e): collapse auth-setup project into globalSetup; parallelize MFA
- replace the 'auth setup' Playwright project (a separate worker-pool + barrier that
  launched its own browsers) with a globalSetup that writes the per-role storageState
  JSON directly (auth is faked — just a cookie) and warms routes once with one browser.
  Removes a whole project phase; on CPU-bound CI this avoids a second worker-pool launch.
- drop mode:'serial' on the MFA describe — its 3 tests are independent read-only page
  checks, so they parallelize across workers (and a failure no longer skips the rest).
